+++ spec/client/user_client_spec.rb
@@ -186,6 +186,39 @@ RSpec.describe KeycloakAdmin::TokenClient do
     end
   end
 
+  describe "#list" do
+    let(:realm_name) { "valid-realm" }
+    let(:user) { KeycloakAdmin::UserRepresentation.from_hash(
+      "username" => "test_username",
+      "createdTimestamp" => Time.now.to_i,
+    )}
+
+    before(:each) do
+      @user_client = KeycloakAdmin.realm(realm_name).users
+
+      stub_token_client
+      allow_any_instance_of(RestClient::Resource).to receive(:get).and_return '[{"username":"test_username","createdTimestamp":1559347200}]'
+    end
+
+    it "lists users" do
+      users = @user_client.list
+      expect(users.length).to eq 1
+      expect(users[0].username).to eq "test_username"
+    end
+
+    it "passes rest client options" do
+      rest_client_options = {verify_ssl: OpenSSL::SSL::VERIFY_NONE}
+      allow_any_instance_of(KeycloakAdmin::Configuration).to receive(:rest_client_options).and_return rest_client_options
+
+      expect(RestClient::Resource).to receive(:new).with(
+        "http://auth.service.io/auth/admin/realms/valid-realm/users", rest_client_options).and_call_original
+
+      users = @user_client.list
+      expect(users.length).to eq 1
+      expect(users[0].username).to eq "test_username"
+    end
+  end
+
   describe "#delete" do
     let(:realm_name) { "valid-realm" }