+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.x.y] - Next version
+
+* List users
+* Support passing rest client options for user save and search
+* Support using gem without Rails/Activerecord
+
 ## [0.7.0] - 2019-06-06
 
 Thanks to @vlad-ro